chore(ui): Rename navigation-menu-premitive.tsx to navigation-menu.tsx
Damit das shadcn Update funktioniert wurde die Navigation Menu Datei umbenannt
